extract interfaces to own files
