+++ python/deps/untypy/untypy/impl/interfaces/iterable.py
@@ -1,0 +1,8 @@
+from typing import Generic, TypeVar, Iterator
+
+I = TypeVar("I")
+
+
+class WIterable(Generic[I]):
+    def __iter__(self) -> Iterator[I]:
+        pass
